Walker executes only if an injection is required (configuration is checked before executing).
