+++ PropertyChanged.Fody/OnChangedWalker.cs
@@ -21,6 +21,8 @@ public partial class ModuleWeaver
     
     IEnumerable<OnChangedMethod> GetOnChangedMethods(TypeNode notifyNode)
     {
+        if (!InjectOnPropertyNameChanged || !InjectImplicitOnPropertyNameChanged)
+            yield break;
         var methods = notifyNode.TypeDefinition.Methods;
 
         var onChangedMethods = methods.Where(x => !x.IsStatic &&
@@ -53,55 +55,6 @@ public partial class ModuleWeaver
                     MethodReference = GetMethodReference(typeDefinitions, methodDefinition)
                 };
             }            
-        }
-    }
-
-    IEnumerable<ExplicitOnPropertyChangedMethodDependency> GetExplicitOnPropertyChangedMethodDependencies(TypeNode notifyNode)
-    {
-        var properties = notifyNode.TypeDefinition.Properties.ToList();
-        var methods = notifyNode.TypeDefinition.Methods;
-
-        foreach (var method in methods.Where(x => !x.IsStatic))
-        {
-            var onPropertyChangedAttribute =
-                method.CustomAttributes.GetAttribute("PropertyChanged.OnPropertyChangedAttribute");
-            if (onPropertyChangedAttribute == null)
-                continue;
-            
-            if (method.ReturnType.FullName != "System.Void")
-            {
-                var message = string.Format("The type {0} has a method with an OnPropertyChanged attribute ({1}) that has a non void return value. Please make the return type void.", notifyNode.TypeDefinition.FullName, method.Name);
-                throw new WeavingException(message);
-            }
-            
-            var propertyDefinitions = GetOnPropertyChangedAttributeProperties(onPropertyChangedAttribute, method, properties);
-
-            OnChangedTypes onChangedType;
-            if (IsNoArgOnChangedMethod(method))
-            {
-                onChangedType = OnChangedTypes.NoArg;
-            }
-            else if (IsBeforeAfterOnChangedMethod(method))
-            {
-                onChangedType = OnChangedTypes.BeforeAfter;
-            }
-            else
-            {
-                var message = string.Format("The type {0} has a method with an OnPropertyChanged attribute ({1}) that has invalid parameter types. Only parameterless methods and methods with two System.Object parameters are supported.", notifyNode.TypeDefinition.FullName, method.Name);
-                throw new WeavingException(message);
-            }
-            foreach (var propertyDefinition in propertyDefinitions)
-            {
-                var typeDefinitions = new Stack<TypeDefinition>();
-                typeDefinitions.Push(notifyNode.TypeDefinition);
-
-                yield return new ExplicitOnPropertyChangedMethodDependency
-                {
-                    OnChangedType = onChangedType,
-                    WhenPropertyIsSet = propertyDefinition,
-                    ShouldCallMethod = GetMethodReference(typeDefinitions, method),
-                };
-            }
         }
     }
